requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
clip==1.0
cycler==0.11.0
dataclasses==0.6
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
fastjsonschema==2.15.3
filelock==3.6.0
fonttools==4.32.0
ftfy==6.1.1
future==0.18.2
gdown==4.4.0
idna==3.3
imageio==2.16.2
importlib-metadata==4.11.3
importlib-resources==5.7.0
ipykernel==6.13.0
ipython==7.32.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.1
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-core==4.9.2
jupyterlab-pygments==0.2.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==2.6.3
ninja==1.10.2.3
notebook==6.4.10
numpy==1.21.6
opencv-python=4.5.5.64
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pip==22.0.4
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.8
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
PyWavelets==1.3.0
pyzmq==22.3.0
regex==2022.3.15
requests==2.27.1
rsa==4.8
scikit-image==0.19.2
scipy==1.7.3
Send2Trash==1.8.0
setuptools==60.10.0
six==1.16.0
soupsieve==2.3.2.post1
terminado==0.13.3
tifffile==2021.11.2
tinycss2==1.1.1
torch==1.7.0+cu101
torchaudio==0.7.0
torchvision==0.8.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.37.1
zipp==3.8.0